# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Fem-ho
  version: 0.1.0
  description: |
    Contracte de l'API de Fem-ho. Aquest fitxer és la font de veritat (regla 5
    d'instruccions.md): els tipus de TypeScript i el client de Kotlin es generen des
    d'aquí i cap client escriu tipus a mà. Un endpoint que no hi és no existeix.

    A M1 només hi ha les dues rutes públiques de vida i identitat. La resta arriba a
    la seva fita, seguint docs/05-api-rest.md.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
servers:
  - url: /api/v1
    description: Base de l'API. GET /info i GET /healthz pengen de l'arrel, no d'aquí.
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Entrar amb correu i contrasenya
      description: |
        Retorna un token d'accés de vida curta i un de refresc de vida llarga.

        **Mai es diu si el correu existeix o no** (docs/02 §2): un correu desconegut i
        una contrasenya errònia donen exactament la mateixa resposta i triguen el
        mateix. Amb 10 intents fallits en 15 minuts comença un bloqueig progressiu
        comptat per correu, no per IP — darrere d'un proxy casolà totes les peticions
        comparteixen IP (docs/05 §7).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Credencials correctes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Correu o contrasenya incorrectes.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          description: Massa intents. Porta `Retry-After`.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refresh
      summary: Rotar el token de refresc
      description: |
        Emet un token nou i invalida l'anterior. Si arriba un token ja gastat es revoca
        tota la família de sessions: és senyal de robatori i no se sap quina de les dues
        bandes és la legítima (docs/05 §1).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token rotat.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Token no vàlid, caducat, revocat o ja gastat.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Tancar la sessió
      responses:
        '204':
          description: Sessió revocada.
        '401':
          description: Cal estar autenticat.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /auth/me:
    get:
      tags: [auth]
      operationId: getMe
      summary: Usuari, capacitats efectives i àmbits
      description: |
        Diu QUÈ pot fer aquesta credencial i A QUINS àmbits arriba. És l'equivalent REST
        de la tool `whoami` d'MCP, i existeix pel mateix motiu: que qui la fa servir
        sàpiga els seus límits abans de trobar-se un 403.
      responses:
        '200':
          description: Qui pregunta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
        '401':
          description: Cal estar autenticat.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  # CONVENCIÓ DE LLISTES
  #
  # Les llistes **paginades** tornen `{ data, next_cursor, has_more }` (docs/05 §3). Les
  # que no es paginen —àmbits, projectes, calendaris— tornen l'array pelat: un envoltori
  # amb un sol camp que no porta cap cursor només afegeix una indirecció a cada client.
  #
  # `/tokens`, `/shares` i `/activity` fan servir l'envoltori sense paginar. És una
  # inconsistència coneguda i es deixa estar a posta: canviar-los ara trencaria els
  # clients per un guany estètic.
    patch:
      tags: [auth]
      operationId: updateProfile
      summary: Modificar el propi perfil
      description: |
        Aquí es toca NOMÉS el compte propi. El brief és explícit (línia 42): al Perfil no
        s'editen els altres; això és a `/admin/users`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                timezone: { type: string, description: 'Nom IANA. Es valida amb `Intl`.' }
                locale:
                  type: string
                  enum: [ca, en, es]
                theme: { type: string, enum: [system, light, dark] }
                accent: { type: string, enum: [default, soft, mono-warm, mono-cool] }
                avatar_color: { type: [string, 'null'] }
      responses:
        '200':
          description: Modificat.
          content: { application/json: { schema: { $ref: '#/components/schemas/UserProfile' } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '422':
          description: Un valor d'enum o un fus que no existeixen.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
  /push/subscriptions:
    post:
      tags: [notifications]
      operationId: subscribePush
      summary: Registrar una subscripció de push
      description: |
        **La mateixa ruta per a Web Push i per a UnifiedPush**: comparteixen RFC i
        xifratge (docs/11 §1). El mateix `endpoint` actualitza la fila en comptes de
        crear-ne una de nova.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint, p256dh, auth, platform]
              properties:
                endpoint: { type: string }
                p256dh: { type: string }
                auth: { type: string }
                platform: { type: string, enum: [web, android] }
      responses:
        '201':
          description: Registrada.
        '401':
          $ref: '#/components/responses/Unauthenticated'
    delete:
      tags: [notifications]
      operationId: unsubscribePush
      summary: Treure una subscripció
      parameters:
        - name: endpoint
          in: query
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Treta.
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /push/public-key:
    get:
      tags: [notifications]
      operationId: getVapidPublicKey
      summary: La clau pública VAPID de la instància
      description: |
        És estable per sempre: es genera un sol cop i **no hi ha rotació** (docs/11 §2).
      responses:
        '200':
          description: La clau.
          content:
            application/json:
              schema:
                type: object
                required: [public_key]
                properties:
                  public_key: { type: string }
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, name, role, kind, timezone, locale, theme, accent]
      properties:
        id: { type: string }
        email: { type: [string, 'null'] }
        name: { type: string }
        role: { type: string, enum: [admin, member] }
        kind: { type: string, enum: [human, ai, caldav_only] }
        timezone: { type: string }
        locale:
          type: string
          enum: [ca, en, es]
          description: |
            L'idioma d'aquesta persona. La primera vegada el client el negocia del
            navegador o del dispositiu i l'escriu aquí; a partir d'aquí el perfil mana
            sobre el dispositiu, que és el que fa que canviar-lo al portàtil també el
            canviï al telèfon.

            **El servidor el fa servir per a les notificacions**, que les pinta el
            sistema operatiu i el client no pot traduir després.
        theme: { type: string, enum: [system, light, dark] }
        accent: { type: string, enum: [default, soft, mono-warm, mono-cool] }
        avatar_color: { type: [string, 'null'] }
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    AuthTokens:
      type: object
      additionalProperties: false
      required: [access_token, refresh_token, expires_at]
      properties:
        access_token:
          type: string
          description: Vida curta. A la web viatja en galeta HttpOnly; a Android com a Bearer.
        refresh_token:
          type: string
          description: Vida llarga i d'un sol ús. Cada refresc n'emet un de nou.
        expires_at:
          type: string
          format: date-time
          description: Caducitat del token de refresc.
    Me:
      type: object
      additionalProperties: false
      required: [id, name, kind, capabilities, scope_ids]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: [string, 'null']
        kind:
          type: string
          enum: [user, agent, guest]
          description: Quina mena de principal és aquesta credencial.
        role:
          type: string
          enum: [admin, member]
        timezone:
          type: string
          description: Les consultes de "avui" es resolen en el fus de qui mira (docs/01 §8).
        capabilities:
          type: array
          items:
            type: string
          description: Les capacitats EFECTIVES, ja intersecades amb el rol del propietari.
        scope_ids:
          type: [array, 'null']
          items:
            type: string
          description: |
            Els àmbits accessibles, o `null` per a tots els del propietari. Regla 9: això
            surt del registre del token, no de cap scope d'OAuth.
        agent_id:
          type: [string, 'null']
          description: Si la credencial és d'un agent, quin. Delegar no és assignar (D5).
...
    Problem:
      type: object
      description: |
        Error en application/problem+json (RFC 9457, docs/05 §3).

        **Qui llegeix un error no és qui el pateix.** El reben tres públics molt
        diferents: una persona davant de l'app, un agent per MCP, i un client CalDAV com
        DAVx⁵ o Thunderbird. Els dos últims no tenen ni catàleg ni idioma. Per això
        l'error viatja en dues peces:

        - `type`, `title` i `detail` **en anglès**, estables, per programar-hi.
        - `params`, les **dades** de l'error. Les apps hi posen el text del seu catàleg
          amb la clau `error.<slug>`, on el `slug` és l'últim tram del `type`. Si no en
          tenen la clau, ensenyen el `detail`: així un error nou del servidor no deixa
          mai una pantalla muda ni obliga a desplegar les dues bandes alhora.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          examples: ['https://femho.app/errors/scope-forbidden']
        title:
          type: string
          examples: ['Scope not accessible']
        params:
          type: object
          additionalProperties: { type: [string, integer] }
          description: |
            Les dades de l'error, planes. Plans a posta: el client no ha de saber
            recórrer un JSON dins d'un JSON per pintar un missatge.
          examples: [{ name: 'Família', count: 3 }]
        status:
          type: integer
          examples: [403]
        detail:
          type: string
          examples: ["Aquest token només té accés a l'àmbit Feina."]
        instance:
          type: string
  responses:
    Unauthenticated:
      description: Cal autenticar-se.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: femho_session
      description: Sessió de navegador. Galeta HttpOnly, Secure, SameSite=Lax.
    bearer:
      type: http
      scheme: bearer
      description: |
        Token d'accés (Android) o token d'API `femho_pat_…` (integracions, MCP, IA).
        Les capacitats i els àmbits van al registre del token, no a scopes d'OAuth
        (regla 9).
security:
  - session: []
  - bearer: []